# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movia Immo - API de gestion immobilière
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{id}/payments:
    get:
      operationId: getTenantPayments
      tags: [tenants]
      summary: Get payments for a specific tenant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # Payments
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: month
          schema:
            type: string
        - in: query
          name: tenantId
          schema:
            type: integer
        - in: query
          name: propertyId
          schema:
            type: integer
      responses:
        "200":
          description: Payment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tenantId
          schema:
            type: integer
        - in: query
          name: propertyId
          schema:
            type: integer
      responses:
        "200":
          description: Contract list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    put:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
  /maintenance:
    get:
      operationId: listMaintenanceTickets
      tags: [maintenance]
      summary: List maintenance tickets
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: propertyId
          schema:
            type: integer
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceTicket"
    post:
      operationId: createMaintenanceTicket
      tags: [maintenance]
      summary: Create a maintenance ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceTicketBody"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
  /maintenance/{id}:
    get:
      operationId: getMaintenanceTicket
      tags: [maintenance]
      summary: Get a maintenance ticket by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
    put:
      operationId: updateMaintenanceTicket
      tags: [maintenance]
      summary: Update a maintenance ticket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        propertyId:
          type: integer
        propertyTitle:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
          enum: [paye, en_retard, partiel, en_attente, annule]
        paymentMethod:
          type: string
          enum: [virement, cheque, especes, mobile_payment]
        penaltyAmount:
          type: number
        receiptNumber:
          type: string
        notes:
          type: string
        month:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reference
        - tenantId
        - propertyId
        - amount
        - dueDate
        - status
        - month
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        tenantId:
          type: integer
        propertyId:
          type: integer
        amount:
          type: number
        paidAmount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
        paymentMethod:
          type: string
        penaltyAmount:
          type: number
        notes:
          type: string
        month:
          type: string
      required:
        - tenantId
        - propertyId
        - amount
        - dueDate
        - status
        - month
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        propertyId:
          type: integer
        propertyTitle:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        chargesAmount:
          type: number
        depositAmount:
          type: number
        depositPaid:
          type: boolean
        type:
          type: string
          enum: [bail_habitation, bail_commercial, saisonnier, courte_duree]
        status:
          type: string
          enum: [actif, expire, resilie, renouvele, en_attente]
        renewalNoticeDate:
          type: string
        specialConditions:
          type: string
        witnessName:
          type: string
        witnessPhone:
          type: string
        daysUntilExpiry:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - reference
        - tenantId
        - propertyId
        - startDate
        - endDate
        - rentAmount
        - type
        - status
        - createdAt
    CreateContractBody:
      type: object
      properties:
        tenantId:
          type: integer
        propertyId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        chargesAmount:
          type: number
        depositAmount:
          type: number
        depositPaid:
          type: boolean
        type:
          type: string
        status:
          type: string
        specialConditions:
          type: string
        witnessName:
          type: string
        witnessPhone:
          type: string
      required:
        - tenantId
        - propertyId
        - startDate
        - endDate
        - rentAmount
        - type
        - status
    MaintenanceTicket:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        title:
          type: string
        description:
          type: string
        propertyId:
          type: integer
        propertyTitle:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        category:
          type: string
          enum: [plomberie, electricite, peinture, menuiserie, climatisation, nettoyage, serrurerie, autre]
        priority:
          type: string
          enum: [faible, normale, haute, urgente]
        status:
          type: string
          enum: [ouvert, en_cours, en_attente_pieces, resolu, ferme]
        technicianName:
          type: string
        technicianPhone:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: number
        scheduledDate:
          type: string
        completedDate:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reference
        - title
        - description
        - propertyId
        - category
        - priority
        - status
        - createdAt
    CreateMaintenanceTicketBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        propertyId:
          type: integer
        tenantId:
          type: integer
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        technicianName:
          type: string
        technicianPhone:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: number
        scheduledDate:
          type: string
        completedDate:
          type: string
        notes:
          type: string
      required:
        - title
        - description
        - propertyId
        - category
        - priority
        - status
...